// src/js/types/WeatherElements.ts
export interface WeatherElements {
  currentTemperature: HTMLDivElement | null;
  currentWeather: HTMLDivElement | null;
  realFeel: HTMLSpanElement | null;
  wind: HTMLSpanElement | null;
  pressure: HTMLSpanElement | null;
  humidity: HTMLSpanElement | null;
  sunrise: HTMLSpanElement | null;
  sunset: HTMLSpanElement | null;
  icon: HTMLElement | null;
}